bot/requirements.txt
fastapi==0.108.0
pydantic==2.5.3
langchain_core==0.1.28
langchain_openai==0.0.5
langchain_community==0.0.25
langchain==0.1.10
requests==2.32.3
google-search-results==2.4.2
pytest==8.3.5
beautifulsoup4==4.12.3
lxml==5.2.2
fake-useragent==1.5.1
sentence-transformers==3.0.1
redis
qdrant_client==1.7.1
uvicorn==0.23.2
python-dotenv==1.0.1